Move Http3Handler into a  separate file.
