+++ neqo-http3/src/lib.rs
@@ -9,6 +9,7 @@
 mod client_events;
 mod connection;
 pub mod connection_client;
+mod connection_server;
 mod control_stream_local;
 mod control_stream_remote;
 pub mod hframe;
